Fixed a number of bugs and missing code, and reworked MultiSelect.
This is not good.
Tried to untie this from dojox/mobile, but it is too deep to do that quickly. The intention was to base this on dojox/mobile/List, not to include it.
Bug in setting options fixed: the options weren't set.
Removed array from prototype. That is not something we want shared.
Added missing destroy and destroyDescendants methods.
Fixed some other bugs. Made code more readable.
Added possibility for a custom sort function.

Changes are bugfixes, and a new backward compatible feature. Version bump.