only remove preloader from DOM on first event, and do that after a wait
While trying to work around a bug in Chrome since v58, cleanup dissolvePreloader

The wait is necessary, because visually it is clear that the transition event comes too soon. We see a jolt. By waiting a while, the dissolve is complete before we yank the preloader element away. This also means user interaction is impossible for 1 more second.